docs/requirements.txt
graphviz
mkdocs
mkdocs-ezglossary-plugin
mkdocs-gen-files
mkdocs-include-markdown-plugin
mkdocs-literate-nav
mkdocs-macros-plugin
mkdocs-mermaid2-plugin
mkdocs-section-index
mkdocs-spellcheck[codespell,symspellpy]
mkdocstrings[python]
pymdown-extensions
